+++ lib/ReactViews/Map/Panels/HelpPanel/HelpPanel.jsx
@@ -1,4 +1,3 @@
-import classNames from "classnames";
 import { observer } from "mobx-react";
 import PropTypes from "prop-types";
 import React from "react";
@@ -6,13 +5,12 @@ import { runInAction } from "mobx";
 import { withTranslation } from "react-i18next";
 import { withTheme } from "styled-components";
 import Icon, { StyledIcon } from "../../../Icon.jsx";
-import Styles from "./help-panel.scss";
 import Spacing from "../../../../Styled/Spacing";
 import Text from "../../../../Styled/Text";
 import Box from "../../../../Styled/Box";
 import parseCustomMarkdownToReact from "../../../Custom/parseCustomMarkdownToReact";
 import HelpPanelItem from "./HelpPanelItem";
-import { RawButton } from "../../../../Styled/Button.jsx";
+import Button, { RawButton } from "../../../../Styled/Button.jsx";
 
 @observer
 class HelpPanel extends React.Component {
@@ -31,36 +29,29 @@ class HelpPanel extends React.Component {
 
   render() {
     const { t } = this.props;
-    // const isVisible = this.props.viewState.showHelpMenu;
     const helpItems = this.props.terria.configParameters.helpContent;
     const isVisible =
       this.props.viewState.showHelpMenu &&
       this.props.viewState.topElement === "HelpPanel";
     const isExpanded = this.props.viewState.helpPanelExpanded;
-    const className = classNames(
-      {
-        [Styles.helpPanel]: true,
-        [Styles.isVisible]: isVisible && !isExpanded,
-        [Styles.isHidden]: !isVisible,
-        [Styles.helpPanelShifted]: isVisible && isExpanded
-      },
-      this.props.viewState.topElement === "HelpPanel" ? "top-element" : ""
-    );
     return (
-      <div
-        className={className}
+      <Box
+        displayInlineBlock
+        backgroundColor={this.props.theme.textLight}
+        styledWidth={"320px"}
+        fullHeight
         onClick={() => this.props.viewState.setTopElement("HelpPanel")}
+        css={`
+          position: fixed;
+          z-index: ${this.props.viewState.topElement === "HelpPanel"
+            ? 99999
+            : 110};
+          transition: right 0.25s;
+          transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+          right: ${isVisible ? (isExpanded ? 490 : 0) : -320}px;
+        `}
       >
-        <div
-          css={`
-            button {
-              padding: 15px;
-              position: absolute;
-              right: 0;
-              z-index: 110;
-            }
-          `}
-        >
+        <Box positionAbsolute paddedRatio={3} topRight>
           <RawButton onClick={() => this.props.viewState.hideHelpPanel()}>
             <StyledIcon
               styledWidth={"16px"}
@@ -69,7 +60,7 @@ class HelpPanel extends React.Component {
               glyph={Icon.GLYPHS.closeLight}
             />
           </RawButton>
-        </div>
+        </Box>
         <Box
           centered
           paddedHorizontally={5}
@@ -90,22 +81,32 @@ class HelpPanel extends React.Component {
           </Text>
           <Spacing bottom={5} />
           <Box centered>
-            <button
-              className={Styles.tourBtn}
-              title={"Take the tour"}
+            <Button
+              primary
+              rounded
+              styledMinWidth={"240px"}
               onClick={() => {
                 runInAction(() => {
                   this.props.viewState.hideHelpPanel();
                   this.props.viewState.setTourIndex(0);
                 });
               }}
+              renderIcon={() => (
+                <StyledIcon
+                  light
+                  styledWidth={"18px"}
+                  glyph={Icon.GLYPHS.tour}
+                />
+              )}
+              textProps={{
+                large: true
+              }}
               css={`
                 ${p => p.theme.addTerriaPrimaryBtnStyles(p)}
               `}
             >
-              {" "}
-              <Icon glyph={Icon.GLYPHS.tour} /> {"Take the tour"}{" "}
-            </button>
+              {"Take the tour"}
+            </Button>
           </Box>
         </Box>
         <Spacing bottom={10} />
@@ -123,7 +124,7 @@ class HelpPanel extends React.Component {
             )}
           </Box>
         </Box>
-      </div>
+      </Box>
     );
   }
 }